refactor(otel): rename plugin.py to invocation_plugin.py
The module holds InvocationOtelPlugin, so name it to match (mirrors
execution_plugin.py). Updated the four importers; the core SDK
aws_durable_execution_sdk_python.plugin module is untouched.
